Title: 길이 단위 변환하기, Time: 6ms, Memory: 0MB, Status: Passed - CodeTree

+++ 240105/길이 단위 변환하기/change-length-unit.cpp
@@ -3,7 +3,7 @@ using namespace std;
 
 int main() {
     double ft = 30.48;
-    float mi = 160934;
+    double mi = 160934;
 
     cout << fixed;
     cout.precision(1);
